pyspark/acm-classifiers/acm-mult-clsf/requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
idna==2.7
Jinja2==2.10
kafka-python==1.4.3
kiwisolver==1.0.1
lightning-python==1.2.1
MarkupSafe==1.0
matplotlib==2.2.2
more-itertools==4.2.0
numpy==1.14.5
pluggy==0.6.0
py==1.5.4
py4j==0.10.7
pyparsing==2.2.0
pyspark==2.3.1
pytest==3.6.3
python-crontab==2.3.4
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
urllib3==1.23
